refactor: group by sub-domain (llm / taunt / result / shared)
