+++ backend/src/features/dialogue/llm/render.ts
