+++ backend/src/features/dialogue/shared/best-effort.ts
